refactor: update NumericFormat component imports and usage across multiple files
