+++ src/components/User/Detail/Crud.tsx
@@ -1,20 +1,16 @@
-import { forwardRef } from 'react'
+import type { HTMLAttributes } from 'react'
 import { FormDataProvider } from '@/providers/FormData'
 import UserDetailCollapsibleCard from './collapsible-card'
 import UserDetailDialogFormWithButton from './DialogFormWithButton'
 
-const UserDetailCrud = forwardRef<HTMLDivElement>(
-    function HarusDikasihNama(props, ref) {
-        return (
-            <div ref={ref} {...props}>
-                <FormDataProvider>
-                    <UserDetailCollapsibleCard
-                        editButton={<UserDetailDialogFormWithButton />}
-                    />
-                </FormDataProvider>
-            </div>
-        )
-    },
-)
-
-export { UserDetailCrud }
+export default function UserDetailCrud(props: HTMLAttributes<HTMLDivElement>) {
+    return (
+        <div {...props}>
+            <FormDataProvider>
+                <UserDetailCollapsibleCard
+                    editButton={<UserDetailDialogFormWithButton />}
+                />
+            </FormDataProvider>
+        </div>
+    )
+}
